backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# AI and Vector Database
google-genai>=0.3.0
qdrant-client==1.16.1
websockets==14.2

# File Processing
pypdf==5.1.0
python-docx==1.1.0
python-pptx==0.6.23

# Utilities
tenacity==8.2.3
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
